Setup package structure. All code is in the pysps/ module and a setup.py file included.
The only sticky part is the buildext.py script needed to carry out the f2py
compilation. Ideally this should be part of setup.py, but the numpy distutils
documentation is rather opaque. The new build setup does not require us
to manually copy fsps/src; that is done automatically in the build script.
All that's required is to run
	python buildext.py
and
	python setup.py develop

to de-facto install the project in site-packages.

Also fixed a compile error in fsps.f90 in the new get_isochrone() function